Rename drink_in_the_fridge to drinks_in_the_fridge
Programme using python while loops to work out how many drinks are left in the fridge. Programme asks user how many types of drinks are in the fridge, how many they have taken and returns how many are left. This has scope to add a return that warns user when they are getting low and to restock.